PP-428: Separete relay hub test into another file
